% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscalingplans_operations.R
\name{create_scaling_plan}
\alias{create_scaling_plan}
\title{Creates a scaling plan}
\usage{
create_scaling_plan(ScalingPlanName, ApplicationSource, ScalingInstructions)
}
\arguments{
\item{ScalingPlanName}{The name of the scaling plan. Names cannot contain vertical bars, colons, or forward slashes.}

\item{ApplicationSource}{A CloudFormation stack or set of tags. You can create one scaling plan per application source.}

\item{ScalingInstructions}{The scaling instructions.}
}
\description{
Creates a scaling plan.
}
\details{
A scaling plan contains a set of instructions used to configure dynamic scaling for the scalable resources in your application. AWS Auto Scaling creates target tracking scaling policies based on the scaling instructions in your scaling plan.
}
\examples{

}
