% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscalingplans_package.R
\docType{package}
\name{paws.autoscalingplans-package}
\alias{paws.autoscalingplans}
\alias{paws.autoscalingplans-package}
\title{paws.autoscalingplans: AWS Auto Scaling Plans}
\description{
Use AWS Auto Scaling to quickly discover all the
scalable AWS resources for your application and configure dynamic
scaling for your scalable resources.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
