% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscalingplans_operations.R
\name{update_scaling_plan}
\alias{update_scaling_plan}
\title{Updates the scaling plan for the specified scaling plan}
\usage{
update_scaling_plan(ScalingPlanName, ScalingPlanVersion,
  ApplicationSource = NULL, ScalingInstructions = NULL)
}
\arguments{
\item{ScalingPlanName}{The name of the scaling plan.}

\item{ScalingPlanVersion}{The version number.}

\item{ApplicationSource}{A CloudFormation stack or set of tags.}

\item{ScalingInstructions}{The scaling instructions.}
}
\description{
Updates the scaling plan for the specified scaling plan.
}
\details{
You cannot update a scaling plan if it is in the process of being created, updated, or deleted.
}
\examples{

}
