% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscalingplans_operations.R
\name{delete_scaling_plan}
\alias{delete_scaling_plan}
\title{Deletes the specified scaling plan}
\usage{
delete_scaling_plan(ScalingPlanName, ScalingPlanVersion)
}
\arguments{
\item{ScalingPlanName}{[required] The name of the scaling plan.}

\item{ScalingPlanVersion}{[required] The version of the scaling plan.}
}
\description{
Deletes the specified scaling plan.
}
\section{Accepted Parameters}{
\preformatted{delete_scaling_plan(
  ScalingPlanName = "string",
  ScalingPlanVersion = 123
)
}
}

