% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscalingplans_operations.R
\name{update_scaling_plan}
\alias{update_scaling_plan}
\title{Updates the scaling plan for the specified scaling plan}
\usage{
update_scaling_plan(ApplicationSource = NULL, ScalingPlanName,
  ScalingInstructions = NULL, ScalingPlanVersion)
}
\arguments{
\item{ApplicationSource}{A CloudFormation stack or set of tags.}

\item{ScalingPlanName}{[required] The name of the scaling plan.}

\item{ScalingInstructions}{The scaling instructions.}

\item{ScalingPlanVersion}{[required] The version number.}
}
\description{
Updates the scaling plan for the specified scaling plan.
}
\details{
You cannot update a scaling plan if it is in the process of being created, updated, or deleted.
}
\section{Accepted Parameters}{
\preformatted{update_scaling_plan(
  ApplicationSource = list(
    CloudFormationStackARN = "string",
    TagFilters = list(
      list(
        Key = "string",
        Values = list(
          "string"
        )
      )
    )
  ),
  ScalingPlanName = "string",
  ScalingInstructions = list(
    list(
      ServiceNamespace = "autoscaling"|"ecs"|"ec2"|"rds"|"dynamodb",
      ResourceId = "string",
      ScalableDimension = "autoscaling:autoScalingGroup:DesiredCapacity"|"ecs:service:DesiredCount"|"ec2:spot-fleet-request:TargetCapacity"|"rds:cluster:ReadReplicaCount"|"dynamodb:table:ReadCapacityUnits"|"dynamodb:table:WriteCapacityUnits"|"dynamodb:index:ReadCapacityUnits"|"dynamodb:index:WriteCapacityUnits",
      MinCapacity = 123,
      MaxCapacity = 123,
      TargetTrackingConfigurations = list(
        list(
          PredefinedScalingMetricSpecification = list(
            PredefinedScalingMetricType = "ASGAverageCPUUtilization"|"ASGAverageNetworkIn"|"ASGAverageNetworkOut"|"DynamoDBReadCapacityUtilization"|"DynamoDBWriteCapacityUtilization"|"ECSServiceAverageCPUUtilization"|"ECSServiceAverageMemoryUtilization"|"ALBRequestCountPerTarget"|"RDSReaderAverageCPUUtilization"|"RDSReaderAverageDatabaseConnections"|"EC2SpotFleetRequestAverageCPUUtilization"|"EC2SpotFleetRequestAverageNetworkIn"|"EC2SpotFleetRequestAverageNetworkOut",
            ResourceLabel = "string"
          ),
          CustomizedScalingMetricSpecification = list(
            MetricName = "string",
            Namespace = "string",
            Dimensions = list(
              list(
                Name = "string",
                Value = "string"
              )
            ),
            Statistic = "Average"|"Minimum"|"Maximum"|"SampleCount"|"Sum",
            Unit = "string"
          ),
          TargetValue = 123.0,
          DisableScaleIn = TRUE|FALSE,
          ScaleOutCooldown = 123,
          ScaleInCooldown = 123,
          EstimatedInstanceWarmup = 123
        )
      )
    )
  ),
  ScalingPlanVersion = 123
)
}
}

